Tidied up a whole heap of crap, mostly methods which had extraneous bracketting. Added in date_display method for lib/authenticated_system.rb which will let users configure how they want the date to be displayed.
The route for locking the topic has now been moved to a PUT http verb instead of a GET, as it suits the purpose of what the action is doing. Also the routes are changed for locking/unlocking topics, removing the dependency of the forum object as that was not really needed.

Also added a theme select box to the user's profile view, this will allow a user to select which theme they want to use on the site. It will only be displayed if there are more than two themes selected.

Removed the USERLEVELS dependency from config/environment.rb and the admin/forums new/edit pages so you can add more userlevels without editing that line, it was stupid anyway.