Changed the width of the post box so the date doesn't line wrap.
Added an else to the edit action of the posts controller so that it doesn't make two tests fail with a rendered or redirected twice
error.

Fixed up the comparison to nil vs fixnum failed in the admin_themes_controller_spec.rb file, this was due to not calling #size on
the @existing_themes object.