Added the rspec and rspec-rails plugin.
Inserted eager-loading for individual forums so that it would load the posts and users for topics inside the paginated query.

Inserted eager-loading for the topic show screen so that it would load the posts and users for the topic.

Added pagination to the topic show screen and the forum show screen (underneath any topic's subject will now be pages if the topic warrants it).

Added in beta searching capabilities using ThinkingSphinx.

Updated the stylesheet to make the pagination links slightly smaller.

Added a migration to add deltas to topics and posts for ThinkingSphinx.

Corrected formatting on the app/views/posts/_post.html.erb