Move route for pages down to the bottom of the routes file so the the lock action for topics actually works.
Pagination links for topics on the forums show screen are now based off the user's per_page value, rather than the hard-coded 30.

The button and drop down to move a set of topics on the forums show screen to another forum will now no longer show if there are no other forums to move it to.

Added 3 indexes to tables to decrease load times. (Topics show now loads up to 5 times faster)