Explicitly autoload the contents of the lib/ folder.
Rails 3 no longer automatically auto-loads the contents of the lib/ folder.
This was changed so that a normal rails application behaves similarly to an engine.

In the future, it will make sense to move some of the contents of the lib/ folder to config/extras/. But that day is not today.