Move the rboard module files from lib/ in to app/lib/
This will guarantee that the module definition(s) will be autoloaded by rails3. It also indicates that the definitions are more tightly integrated with the application itself, rather than being application-independent extensions or other autonomous fragments of code.

Moving the definitions in to app/lib also causes some cucumber tests to pass again.